Move timeline to its own module
