+++ placehero/src/main.rs
@@ -3,7 +3,8 @@
 
 //! The boilerplate run function for desktop platforms
 
-use xilem::{EventLoop, winit::error::EventLoopError};
+use xilem::EventLoop;
+use xilem::winit::error::EventLoopError;
 
 fn main() -> Result<(), EventLoopError> {
     placehero::run(EventLoop::with_user_event())
